Extract `Simulation` class into its own files
